refactor(selectCombo)
