+++ src/es6/content.js
@@ -1,9 +1,8 @@
 import React from "react";
-import ProjectGallery from "./projectGallery.js";
 
 import { AppState } from "./contexts";
-import contextState from "../../state/state";
-import ProjectButton from "./projectButton.js";
+import { SelectCombo } from "./selectCombo";
+import ProjectGallery from "./projectGallery.js";
 
 export class Content extends React.Component {
   static contextType = AppState;
@@ -13,46 +12,6 @@ export class Content extends React.Component {
     ) : (
       <section className="gallery">
         <ProjectGallery lng={this.context.lng} project={this.context.project} />
-      </section>
-    );
-  }
-}
-
-export class SelectCombo extends React.Component {
-  static contextType = AppState;
-  render() {
-    const { projects } = contextState;
-    const lng = this.context.lng || contextState.meta.languageDefault;
-    const section = !this.context.project ? "home" : "project";
-    const className = !this.context.project ? "menu menu-content" : "menu menu-header";
-    let projectSt = this.context.project || {};
-    return (
-      <section className={className}>
-        {" "}
-        {Object.keys(projects)
-          .map(function(key) {
-            return projects[key];
-          })
-          .map(function(project, index) {
-            project.state = project.state || "";
-            const name = !lng ? project.name : project.name[lng];
-
-            return (
-              <ProjectButton
-                icon={project.icon}
-                view={section}
-                lng={lng}
-                key={index}
-                key2={Object.keys(projects)[index]}
-                name={name}
-                color={project.color}
-                type={project.type}
-                publish={project.publish}
-                state={project.state}
-                active={project.name === projectSt.name}
-              />
-            );
-          }, this)}{" "}
       </section>
     );
   }
