+++ src/es6/footer.js
@@ -1,10 +1,10 @@
+import { faGithub } from "@fortawesome/free-brands-svg-icons";
+import { faPhone, faEnvelope } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
 
 import { AppState } from "./contexts";
 import contextState from "../../state/state";
-import { faGithub } from "@fortawesome/free-brands-svg-icons";
-import { faPhone, faEnvelope } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 export class Footer extends React.Component {
   static contextType = AppState;
